[!!!][FEATURE] Respect gutter/correction for image rendering
This feature has an effect only in conjunction with the
template pizpalue (since v11.6.0, introducing `image
variants modifier stack`). Column gutter and correction
are now respected in column structure elements. The
feature is enabled by default which might lead to changed
images in the frontend which under most circumstances is
desired.

**Corrective action:**
Check if images are shown correctly. To render images as
before disable the features by help of the constant
editor. Alternatively add the following to the ts setup:

```
lib.containerContentElement.settings.containerElements {
   imageVariants {
      gutter.enable = 0
      correction.enable = 0
   }
}
```